[eng-1014] Don't delete the "parts" node that other nodes might be already watching for changes
