+++ src/main/java/com/sonian/elasticsearch/zookeeper/discovery/ZooKeeperDiscovery.java
@@ -44,7 +44,6 @@ import com.sonian.elasticsearch.zookeeper.client.AbstractNodeListener;
 
 import java.io.IOException;
 import java.util.HashSet;
-import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.atomic.AtomicBoolean;
